+++ lib/accessibility_tools.dart
@@ -3,3 +3,4 @@ library accessibility_tools;
 export 'src/accessibility_tools.dart';
 export 'src/checkers/minimum_tap_area_checker.dart' show MinimumTapAreas;
 export 'src/enums/buttons_alignment.dart';
+export 'src/testing_tools/testing_tools_configuration.dart';